+++ changelog.md
@@ -1,6 +1,7 @@
 ## Forestry 2.1.0
 - Fixed typo that cut off black order on the right side of the Portable Analyzer GUI
 - Add IActivityType bee chromosome for more complex sleep patterns (#97)
+- Add missing recipes for Stripped Wood, Fireproof Wood, and Fireproof Stripped Wood
 
 ## Forestry 2.0.5
 - Fixed KubeJS server event error (#138)